html/semantics/forms/the-select-element/stylable-select/select-accessibility-minimum-target-size.tentative.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://www.w3.org/WAI/WCAG22/Understanding/target-size-minimum.html">
<link rel=help href="https://github.com/openui/open-ui/issues/1026">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<!-- TODO also test vertical writing modes -->

<select style="appearance:base-select">
  <option>option</option>
</select>

<script>
promise_test(async () => {
  const select = document.querySelector('select');
  const buttonRect = select.getBoundingClientRect();
  assert_greater_than_equal(buttonRect.width, 24,
    'Button width must be at least 24 pixels.');
  assert_greater_than_equal(buttonRect.height, 24,
    'Button height must be at least 24 pixels.');

  await test_driver.bless();
  select.showPicker();
  const optionRect = document.querySelector('option').getBoundingClientRect();
  assert_greater_than_equal(optionRect.width, 24,
    'Option width must be at least 24 pixels.');
  assert_greater_than_equal(optionRect.height, 24,
    'Option height must be at least 24 pixels.');
}, 'Select with appearance:base-select should have targets for pointer inputs be at least 24x24 CSS pixels.');
</script>
